+++ packages/tools/utils/src/[any-any]-[object-array].spec.js
@@ -2,7 +2,10 @@ import {strict as assert} from 'assert';
 
 import * as _ from 'lamb';
 
-import {objectToKeyValueArrayWith} from './[any-any]-[object-array]';
+import {
+	objectToKeyValueArrayWith,
+	valuesWith
+} from './[any-any]-[object-array]';
 
 describe('(Any -> Any) -> (Object -> Array)', function() {
 	describe('objectToKeyValueArrayWith', function() {
@@ -13,6 +16,33 @@ describe('(Any -> Any) -> (Object -> Array)', function() {
 			assert.deepStrictEqual(
 				convertToArray(obj),
 				[{key: 'k1', value: 1}, {key: 'k2', value: 2}]
+			);
+		});
+	});
+	describe('valuesWith', function() {
+		it('should return a function expecting an object and returning an array of its value processed with the provided function', function() {
+			const triplicatedValues = valuesWith(_.add(3));
+
+			assert.deepStrictEqual(
+				triplicatedValues({a: 1, b: 2, c: 3}),
+				[4, 5, 6]
+			);
+
+			const getFoos = valuesWith(_.getKey('foo'));
+			const obj = {
+				a: {foo: 1, bar: 2},
+				b: {foo: 15, bar: -3},
+				c: {foo: 4, bar: 12}
+			};
+
+			assert.deepStrictEqual(getFoos(obj), [1, 15, 4]);
+		});
+		it('should return a function expecting an object and returning an array of its value processed with the provided function - using keys', function() {
+			const keysAndValues = valuesWith((value, key) => `${value} (${key})`);
+
+			assert.deepStrictEqual(
+				keysAndValues({a: 3, b: 5}),
+				['3 (a)', '5 (b)']
 			);
 		});
 	});